tool: move staging area operations from parse-config.c to staging-area.c
